Move MockCallAdapter to lib
